refactor: 연관 도메인으로 디렉토링 이동 및 리팩토링
