+++ helper-persistence/src/main/java/org/example/persistence/student/entity/GradeInfoId.java
@@ -1,4 +1,4 @@
-package org.example.persistence.gradeInfo.entity;
+package org.example.persistence.student.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
