+++ helper-persistence/src/main/java/org/example/persistence/student/mapper/StudentMapper.java
@@ -5,8 +5,8 @@ import org.example.domain.student.model.Student;
 import org.example.persistence.GenericMapper;
 import org.example.persistence.auth.entity.UserJpaEntity;
 import org.example.persistence.auth.repository.UserJpaRepository;
-import org.example.persistence.gradeInfo.entity.GradeInfo;
-import org.example.persistence.gradeInfo.repository.GradeInfoJpaRepository;
+import org.example.persistence.student.entity.GradeInfo;
+import org.example.persistence.student.repository.GradeInfoJpaRepository;
 import org.example.persistence.student.entity.StudentJpaEntity;
 import org.springframework.stereotype.Component;
